Updates functions to let user select data interval
